requirements.txt.txt
pyTelegramBotAPI
pydub
SpeechRecognition
requests
PyJWT
